CREATED DAO.KT

+++ src/main/kotlin/Tienda.kt
@@ -1,1 +1,1 @@
-data class Tienda(val id:Int, val nombre:String, val direccion:String)
+data class Tienda(val id:Long, val nombre:String, val direccion:String)
